+++ CHANGELOG.md
@@ -7,7 +7,12 @@ title: Changelog
 ### Bug Fixes
 
 -   Add special handling for import types with type errors discarded with ts-expect-error, #2792.
+-   Improved support for hosting TypeDoc with strict Content Security Policy rules, #2794.
 -   Fixed low contrast in default colors for properties/accessors in light mode, #2795.
+
+### Thanks!
+
+-   @mikalai-snap
 
 ## v0.27.2 (2024-11-29)